js/lite: rename origin-id module to origin
The helper only exports randomOriginId, so the shorter name fits and
follows the file-per-concept convention used elsewhere in lite/.
