+++ js/lite/src/lite/origin.ts
